+++ Sources/StreamVideo/Controllers/CallController.swift
@@ -495,6 +495,10 @@ class CallController: @unchecked Sendable {
             .sink { [weak self] in self?.webRTCClientDidUpdateStage($0) }
     }
 
+    internal func callKitActivated(_ audioSession: AVAudioSessionProtocol) throws {
+        try webRTCCoordinator.callKitActivated(audioSession)
+    }
+
     // MARK: - private
 
     private func handleParticipantsUpdated() {